+++ source/production/F0.Generators/Diagnostics/DiagnosticCategories.cs
@@ -1,0 +1,8 @@
+namespace F0.Diagnostics;
+
+internal static class DiagnosticCategories
+{
+	internal const string SourceGenerationExceptionGenerator = nameof(SourceGenerationExceptionGenerator); //F0GEN01xx
+	internal const string FriendlyNameGenerator = nameof(FriendlyNameGenerator); //F0GEN02xx
+	internal const string EnumInfoGenerator = nameof(EnumInfoGenerator); //F0GEN03xx
+}