+++ source/test/F0.Generators.Tests/Shared/SourceGenerationExceptionGeneratorTests.cs
@@ -1,3 +1,4 @@
+using F0.Diagnostics;
 using F0.Shared;
 using F0.Tests.Generated;
 using F0.Tests.Verifiers;
@@ -49,7 +50,7 @@ public class SourceGenerationExceptionGeneratorTests
 }
 ";
 
-		await VerifyAsync(test, generated);
+		await VerifyAsync(test, null, generated, null);
 	}
 
 	[Fact]
@@ -95,7 +96,7 @@ public class SourceGenerationExceptionGeneratorTests
 }
 ";
 
-		await VerifyAsync(test, generated, LanguageVersion.CSharp7_3);
+		await VerifyAsync(test, null, generated, LanguageVersion.CSharp7_3);
 	}
 
 	[Fact]
@@ -141,14 +142,110 @@ public class SourceGenerationExceptionGeneratorTests
 }
 ";
 
-		await VerifyAsync(test, generated, LanguageVersion.CSharp1);
+		await VerifyAsync(test, null, generated, LanguageVersion.CSharp1);
 	}
 
-	private static Task VerifyAsync(string test, string generated, LanguageVersion? languageVersion = null)
+	[Fact]
+	public async Task Execute_AvoidUsage()
 	{
+		string test =
+@"#nullable enable
+using System;
+using System.Collections.Generic;
+using F0.Generated;
+
+public sealed class Class
+{
+	internal void Method(object obj, {|#0:SourceGenerationException?|} ex)
+	{
+		try
+		{
+			_ = new {|#1:SourceGenerationException|}("""");
+			_ = obj as {|#2:SourceGenerationException|};
+			_ = ({|#3:SourceGenerationException|})obj;
+			_ = new List<{|#4:SourceGenerationException|}>(240);
+		}
+		catch ({|#5:SourceGenerationException|} exception)
+		{
+			throw new {|#6:SourceGenerationException|}("""");
+		}
+		catch (Exception exception) when (exception is {|#7:F0.Generated.SourceGenerationException|})
+		{
+			throw new {|#8:F0.Generated.SourceGenerationException|}("""");
+		}
+	}
+}
+";
+
+		DiagnosticResult[] diagnostics = new[]
+		{
+			CreateDiagnostic(0, "SourceGenerationException? ex"),
+			CreateDiagnostic(1, "new SourceGenerationException(\"\")"),
+			CreateDiagnostic(2, "obj as SourceGenerationException"),
+			CreateDiagnostic(3, "(SourceGenerationException)obj"),
+			CreateDiagnostic(4, "<SourceGenerationException>"),
+			CreateDiagnostic(5, "(SourceGenerationException exception)"),
+			CreateDiagnostic(6, "throw new SourceGenerationException(\"\");"),
+			CreateDiagnostic(7, "exception is F0.Generated.SourceGenerationException"),
+			CreateDiagnostic(8, "throw new F0.Generated.SourceGenerationException(\"\");"),
+		};
+
+		string generated =
+@"namespace F0.Generated
+{
+	internal sealed class SourceGenerationException : global::System.Exception
+	{
+		private const string helpLink = ""https://github.com/Flash0ver/F0.Generators"";
+
+		public SourceGenerationException()
+			: base(CreateNotGeneratedMessage())
+		{
+			HelpLink = helpLink;
+		}
+
+		public SourceGenerationException(string message)
+			: base(message)
+		{
+			HelpLink = helpLink;
+		}
+
+		public SourceGenerationException(string message, global::System.Exception innerException)
+			: base(message, innerException)
+		{
+			HelpLink = helpLink;
+		}
+
+		private static string CreateNotGeneratedMessage()
+		{
+			const string uri = ""https://github.com/Flash0ver/F0.Generators/issues"";
+
+			return ""The method or operation was not generated correctly.""
+				+ "" Please leave a comment on a related issue, or create a new issue at ""
+				+ ""'"" + uri + ""'""
+				+ "". Thank you!"";
+		}
+	}
+}
+";
+
+		await VerifyAsync(test, diagnostics, generated, null);
+	}
+
+	private static DiagnosticResult CreateDiagnostic(int markupKey, string expression)
+	{
+		return CSharpSourceGeneratorVerifier<SourceGenerationExceptionGenerator>.Diagnostic(DiagnosticIds.F0GEN0101, DiagnosticSeverity.Warning)
+			.WithMessageFormat($"Avoid using 'SourceGenerationException' directly: '{{0}}'")
+			.WithArguments(expression)
+			.WithLocation(markupKey);
+	}
+
+	private static Task VerifyAsync(string test, DiagnosticResult[]? diagnostics, string generated, LanguageVersion? languageVersion)
+	{
+		diagnostics ??= Array.Empty<DiagnosticResult>();
+
 		string filename = $@"F0.Generators\{typeof(SourceGenerationExceptionGenerator).FullName}\SourceGenerationException.g.cs";
 		string content = String.Concat(Sources.GetFileHeader(languageVersion), generated);
 
-		return CSharpSourceGeneratorVerifier<SourceGenerationExceptionGenerator>.VerifySourceGeneratorAsync(test, (filename, content), languageVersion);
+		return CSharpSourceGeneratorVerifier<SourceGenerationExceptionGenerator>.VerifySourceGeneratorAsync(test, diagnostics, (filename, content), languageVersion);
 	}
 }